Removes deprecated useAuthToken; Removes deprecated functions from root hub-service; Renames root hub-service to hub-token-service since it only provides the token now.
