+++ cardstack/src/hooks/index.ts
@@ -1,7 +1,6 @@
 export * from './transaction-confirmation/use-transaction-confirmation';
 export * from './transactions';
 export * from './use-lifetime-earnings-data';
-export * from './prepaid-card/useAuthToken';
 export * from './notifications-preferences/useUpdateNotificationPreferences';
 export * from './useMutationEffects';
 export * from './useCopyWithFeedback';
